Consolidate toast code in ContainerActivity
All toasts are now displayed via a single toast object in
ContainerActivity. This fixes a bug that caused the app to crash when
clearing the lexicon history or syntax bookmarks, ensures that we'll
never have overlapping toasts, and simplifies toast creation throughout
the app.
