+++ GreekReference/src/main/java/com/benlinskey/greekreference/AbstractDetailFragment.java
@@ -16,11 +16,9 @@
 
 package com.benlinskey.greekreference;
 
-import android.annotation.SuppressLint;
 import android.app.Fragment;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Toast;
 
 /**
  * The basic class from which every detail fragment inherits.
@@ -28,17 +26,6 @@ import android.widget.Toast;
 public abstract class AbstractDetailFragment extends Fragment {
     
     private static final String KEY_SCROLL_Y = "scroll_y";
-
-    // We use a single Toast object to prevent overlapping toasts when the user
-    // repeatedly taps an icon that displays a toast.
-    protected Toast mToast;
-
-    @SuppressLint("ShowToast")
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        mToast = Toast.makeText(getActivity(), null, Toast.LENGTH_SHORT);
-    }
 
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
@@ -58,15 +45,4 @@ public abstract class AbstractDetailFragment extends Fragment {
         }
     }
 
-    /**
-     * Displays a toast containing the specified text.
-     * <p>
-     * All children of this class should display toasts only by calling this
-     * method in order to prevent creating overlapping toasts.
-     * @param message the text to display in the toast
-     */
-    public void displayToast(String message) {
-        mToast.setText(message);
-        mToast.show();
-    }
 }
